refactoring functions out of main
